+++ changelog.md
@@ -5,6 +5,7 @@
     - temp arenas are pooled per thread. rec arena is exactly one per thread.
       this is effectively how it worked previously too. but now the impls are separated
       and much simpler.
+- `Vec::clone{_in}` now guarantees `clone.len() == clone.cap()`.
 
 0.3.0:
 - add `borrow::{BorrowFlag, BorrowRef, BorrowRefMut, Ref, RefMut}`.